@@ -14,7 +14,7 @@
 
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.27.1
+// 	protoc-gen-go v1.35.2
 // 	protoc        (unknown)
 // source: google/api/annotations.proto
 
@@ -80,7 +80,7 @@ var file_google_api_annotations_proto_rawDesc = []byte{
 	0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
 }
 
-var file_google_api_annotations_proto_goTypes = []interface{}{
+var file_google_api_annotations_proto_goTypes = []any{
 	(*descriptorpb.MethodOptions)(nil), // 0: google.protobuf.MethodOptions
 	(*HttpRule)(nil),                   // 1: google.api.HttpRule
 }
