@@ -5,16 +5,17 @@ import (
 	"fmt"
 	"net/http"
 
-	pb "github.com/anicoll/unicom/gen/pb/go/unicom/api/v1"
 	"github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials/insecure"
+
+	pb "github.com/anicoll/unicom/gen/pb/go/unicom/api/v1"
 )
 
 func runHTTPGateway(ctx context.Context, httpPort, grpcPort int) error {
 	mux := runtime.NewServeMux()
 	opts := []grpc.DialOption{grpc.WithTransportCredentials(insecure.NewCredentials())}
-	err := pb.RegisterUnicomHandlerFromEndpoint(ctx, mux, fmt.Sprintf("localhost:%d", grpcPort), opts)
+	err := pb.RegisterUnicomServiceHandlerFromEndpoint(ctx, mux, fmt.Sprintf("localhost:%d", grpcPort), opts)
 	if err != nil {
 		return err
 	}
