@@ -10,6 +10,7 @@ package apiv1
 
 import (
 	"context"
+	"errors"
 	"io"
 	"net/http"
 
@@ -24,247 +25,205 @@ import (
 )
 
 // Suppress "imported and not used" errors
-var _ codes.Code
-var _ io.Reader
-var _ status.Status
-var _ = runtime.String
-var _ = utilities.NewDoubleArray
-var _ = metadata.Join
+var (
+	_ codes.Code
+	_ io.Reader
+	_ status.Status
+	_ = errors.New
+	_ = runtime.String
+	_ = utilities.NewDoubleArray
+	_ = metadata.Join
+)
 
-func request_Unicom_SendCommunication_0(ctx context.Context, marshaler runtime.Marshaler, client UnicomClient, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
-	var protoReq SendCommunicationRequest
-	var metadata runtime.ServerMetadata
-
-	newReader, berr := utilities.IOReaderFactory(req.Body)
-	if berr != nil {
-		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", berr)
-	}
-	if err := marshaler.NewDecoder(newReader()).Decode(&protoReq); err != nil && err != io.EOF {
+func request_UnicomService_SendCommunication_0(ctx context.Context, marshaler runtime.Marshaler, client UnicomServiceClient, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
+	var (
+		protoReq SendCommunicationRequest
+		metadata runtime.ServerMetadata
+	)
+	if err := marshaler.NewDecoder(req.Body).Decode(&protoReq); err != nil && !errors.Is(err, io.EOF) {
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
 	}
-
 	msg, err := client.SendCommunication(ctx, &protoReq, grpc.Header(&metadata.HeaderMD), grpc.Trailer(&metadata.TrailerMD))
 	return msg, metadata, err
-
 }
 
-func local_request_Unicom_SendCommunication_0(ctx context.Context, marshaler runtime.Marshaler, server UnicomServer, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
-	var protoReq SendCommunicationRequest
-	var metadata runtime.ServerMetadata
-
-	newReader, berr := utilities.IOReaderFactory(req.Body)
-	if berr != nil {
-		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", berr)
-	}
-	if err := marshaler.NewDecoder(newReader()).Decode(&protoReq); err != nil && err != io.EOF {
+func local_request_UnicomService_SendCommunication_0(ctx context.Context, marshaler runtime.Marshaler, server UnicomServiceServer, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
+	var (
+		protoReq SendCommunicationRequest
+		metadata runtime.ServerMetadata
+	)
+	if err := marshaler.NewDecoder(req.Body).Decode(&protoReq); err != nil && !errors.Is(err, io.EOF) {
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
 	}
-
 	msg, err := server.SendCommunication(ctx, &protoReq)
 	return msg, metadata, err
-
 }
 
-func request_Unicom_GetStatus_0(ctx context.Context, marshaler runtime.Marshaler, client UnicomClient, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
-	var protoReq GetStatusRequest
-	var metadata runtime.ServerMetadata
-
+func request_UnicomService_GetStatus_0(ctx context.Context, marshaler runtime.Marshaler, client UnicomServiceClient, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
 	var (
-		val string
-		ok  bool
-		err error
-		_   = err
+		protoReq GetStatusRequest
+		metadata runtime.ServerMetadata
+		err      error
 	)
-
-	val, ok = pathParams["id"]
+	val, ok := pathParams["id"]
 	if !ok {
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "missing parameter %s", "id")
 	}
-
 	protoReq.Id, err = runtime.String(val)
 	if err != nil {
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "type mismatch, parameter: %s, error: %v", "id", err)
 	}
-
 	msg, err := client.GetStatus(ctx, &protoReq, grpc.Header(&metadata.HeaderMD), grpc.Trailer(&metadata.TrailerMD))
 	return msg, metadata, err
-
 }
 
-func local_request_Unicom_GetStatus_0(ctx context.Context, marshaler runtime.Marshaler, server UnicomServer, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
-	var protoReq GetStatusRequest
-	var metadata runtime.ServerMetadata
-
+func local_request_UnicomService_GetStatus_0(ctx context.Context, marshaler runtime.Marshaler, server UnicomServiceServer, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
 	var (
-		val string
-		ok  bool
-		err error
-		_   = err
+		protoReq GetStatusRequest
+		metadata runtime.ServerMetadata
+		err      error
 	)
-
-	val, ok = pathParams["id"]
+	val, ok := pathParams["id"]
 	if !ok {
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "missing parameter %s", "id")
 	}
-
 	protoReq.Id, err = runtime.String(val)
 	if err != nil {
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "type mismatch, parameter: %s, error: %v", "id", err)
 	}
-
 	msg, err := server.GetStatus(ctx, &protoReq)
 	return msg, metadata, err
-
 }
 
-// RegisterUnicomHandlerServer registers the http handlers for service Unicom to "mux".
-// UnaryRPC     :call UnicomServer directly.
+// RegisterUnicomServiceHandlerServer registers the http handlers for service UnicomService to "mux".
+// UnaryRPC     :call UnicomServiceServer directly.
 // StreamingRPC :currently unsupported pending https://github.com/grpc/grpc-go/issues/906.
-// Note that using this registration option will cause many gRPC library features to stop working. Consider using RegisterUnicomHandlerFromEndpoint instead.
-func RegisterUnicomHandlerServer(ctx context.Context, mux *runtime.ServeMux, server UnicomServer) error {
-
-	mux.Handle("POST", pattern_Unicom_SendCommunication_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
+// Note that using this registration option will cause many gRPC library features to stop working. Consider using RegisterUnicomServiceHandlerFromEndpoint instead.
+// GRPC interceptors will not work for this type of registration. To use interceptors, you must use the "runtime.WithMiddlewares" option in the "runtime.NewServeMux" call.
+func RegisterUnicomServiceHandlerServer(ctx context.Context, mux *runtime.ServeMux, server UnicomServiceServer) error {
+	mux.Handle(http.MethodPost, pattern_UnicomService_SendCommunication_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
 		ctx, cancel := context.WithCancel(req.Context())
 		defer cancel()
 		var stream runtime.ServerTransportStream
 		ctx = grpc.NewContextWithServerTransportStream(ctx, &stream)
 		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
-		var err error
-		ctx, err = runtime.AnnotateIncomingContext(ctx, mux, req, "/unicom.api.v1.Unicom/SendCommunication", runtime.WithHTTPPathPattern("/unicom/v1/send-communication"))
+		annotatedContext, err := runtime.AnnotateIncomingContext(ctx, mux, req, "/unicom.api.v1.UnicomService/SendCommunication", runtime.WithHTTPPathPattern("/unicom/v1/send-communication"))
 		if err != nil {
 			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
 			return
 		}
-		resp, md, err := local_request_Unicom_SendCommunication_0(ctx, inboundMarshaler, server, req, pathParams)
+		resp, md, err := local_request_UnicomService_SendCommunication_0(annotatedContext, inboundMarshaler, server, req, pathParams)
 		md.HeaderMD, md.TrailerMD = metadata.Join(md.HeaderMD, stream.Header()), metadata.Join(md.TrailerMD, stream.Trailer())
-		ctx = runtime.NewServerMetadataContext(ctx, md)
+		annotatedContext = runtime.NewServerMetadataContext(annotatedContext, md)
 		if err != nil {
-			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
+			runtime.HTTPError(annotatedContext, mux, outboundMarshaler, w, req, err)
 			return
 		}
-
-		forward_Unicom_SendCommunication_0(ctx, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)
-
+		forward_UnicomService_SendCommunication_0(annotatedContext, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)
 	})
-
-	mux.Handle("GET", pattern_Unicom_GetStatus_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
+	mux.Handle(http.MethodGet, pattern_UnicomService_GetStatus_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
 		ctx, cancel := context.WithCancel(req.Context())
 		defer cancel()
 		var stream runtime.ServerTransportStream
 		ctx = grpc.NewContextWithServerTransportStream(ctx, &stream)
 		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
-		var err error
-		ctx, err = runtime.AnnotateIncomingContext(ctx, mux, req, "/unicom.api.v1.Unicom/GetStatus", runtime.WithHTTPPathPattern("/unicom/v1/status/{id}"))
+		annotatedContext, err := runtime.AnnotateIncomingContext(ctx, mux, req, "/unicom.api.v1.UnicomService/GetStatus", runtime.WithHTTPPathPattern("/unicom/v1/status/{id}"))
 		if err != nil {
 			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
 			return
 		}
-		resp, md, err := local_request_Unicom_GetStatus_0(ctx, inboundMarshaler, server, req, pathParams)
+		resp, md, err := local_request_UnicomService_GetStatus_0(annotatedContext, inboundMarshaler, server, req, pathParams)
 		md.HeaderMD, md.TrailerMD = metadata.Join(md.HeaderMD, stream.Header()), metadata.Join(md.TrailerMD, stream.Trailer())
-		ctx = runtime.NewServerMetadataContext(ctx, md)
+		annotatedContext = runtime.NewServerMetadataContext(annotatedContext, md)
 		if err != nil {
-			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
+			runtime.HTTPError(annotatedContext, mux, outboundMarshaler, w, req, err)
 			return
 		}
-
-		forward_Unicom_GetStatus_0(ctx, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)
-
+		forward_UnicomService_GetStatus_0(annotatedContext, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)
 	})
 
 	return nil
 }
 
-// RegisterUnicomHandlerFromEndpoint is same as RegisterUnicomHandler but
+// RegisterUnicomServiceHandlerFromEndpoint is same as RegisterUnicomServiceHandler but
 // automatically dials to "endpoint" and closes the connection when "ctx" gets done.
-func RegisterUnicomHandlerFromEndpoint(ctx context.Context, mux *runtime.ServeMux, endpoint string, opts []grpc.DialOption) (err error) {
-	conn, err := grpc.Dial(endpoint, opts...)
+func RegisterUnicomServiceHandlerFromEndpoint(ctx context.Context, mux *runtime.ServeMux, endpoint string, opts []grpc.DialOption) (err error) {
+	conn, err := grpc.NewClient(endpoint, opts...)
 	if err != nil {
 		return err
 	}
 	defer func() {
 		if err != nil {
 			if cerr := conn.Close(); cerr != nil {
-				grpclog.Infof("Failed to close conn to %s: %v", endpoint, cerr)
+				grpclog.Errorf("Failed to close conn to %s: %v", endpoint, cerr)
 			}
 			return
 		}
 		go func() {
 			<-ctx.Done()
 			if cerr := conn.Close(); cerr != nil {
-				grpclog.Infof("Failed to close conn to %s: %v", endpoint, cerr)
+				grpclog.Errorf("Failed to close conn to %s: %v", endpoint, cerr)
 			}
 		}()
 	}()
-
-	return RegisterUnicomHandler(ctx, mux, conn)
+	return RegisterUnicomServiceHandler(ctx, mux, conn)
 }
 
-// RegisterUnicomHandler registers the http handlers for service Unicom to "mux".
+// RegisterUnicomServiceHandler registers the http handlers for service UnicomService to "mux".
 // The handlers forward requests to the grpc endpoint over "conn".
-func RegisterUnicomHandler(ctx context.Context, mux *runtime.ServeMux, conn *grpc.ClientConn) error {
-	return RegisterUnicomHandlerClient(ctx, mux, NewUnicomClient(conn))
+func RegisterUnicomServiceHandler(ctx context.Context, mux *runtime.ServeMux, conn *grpc.ClientConn) error {
+	return RegisterUnicomServiceHandlerClient(ctx, mux, NewUnicomServiceClient(conn))
 }
 
-// RegisterUnicomHandlerClient registers the http handlers for service Unicom
-// to "mux". The handlers forward requests to the grpc endpoint over the given implementation of "UnicomClient".
-// Note: the gRPC framework executes interceptors within the gRPC handler. If the passed in "UnicomClient"
+// RegisterUnicomServiceHandlerClient registers the http handlers for service UnicomService
+// to "mux". The handlers forward requests to the grpc endpoint over the given implementation of "UnicomServiceClient".
+// Note: the gRPC framework executes interceptors within the gRPC handler. If the passed in "UnicomServiceClient"
 // doesn't go through the normal gRPC flow (creating a gRPC client etc.) then it will be up to the passed in
-// "UnicomClient" to call the correct interceptors.
-func RegisterUnicomHandlerClient(ctx context.Context, mux *runtime.ServeMux, client UnicomClient) error {
-
-	mux.Handle("POST", pattern_Unicom_SendCommunication_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
+// "UnicomServiceClient" to call the correct interceptors. This client ignores the HTTP middlewares.
+func RegisterUnicomServiceHandlerClient(ctx context.Context, mux *runtime.ServeMux, client UnicomServiceClient) error {
+	mux.Handle(http.MethodPost, pattern_UnicomService_SendCommunication_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
 		ctx, cancel := context.WithCancel(req.Context())
 		defer cancel()
 		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
-		var err error
-		ctx, err = runtime.AnnotateContext(ctx, mux, req, "/unicom.api.v1.Unicom/SendCommunication", runtime.WithHTTPPathPattern("/unicom/v1/send-communication"))
+		annotatedContext, err := runtime.AnnotateContext(ctx, mux, req, "/unicom.api.v1.UnicomService/SendCommunication", runtime.WithHTTPPathPattern("/unicom/v1/send-communication"))
 		if err != nil {
 			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
 			return
 		}
-		resp, md, err := request_Unicom_SendCommunication_0(ctx, inboundMarshaler, client, req, pathParams)
-		ctx = runtime.NewServerMetadataContext(ctx, md)
+		resp, md, err := request_UnicomService_SendCommunication_0(annotatedContext, inboundMarshaler, client, req, pathParams)
+		annotatedContext = runtime.NewServerMetadataContext(annotatedContext, md)
 		if err != nil {
-			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
+			runtime.HTTPError(annotatedContext, mux, outboundMarshaler, w, req, err)
 			return
 		}
-
-		forward_Unicom_SendCommunication_0(ctx, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)
-
+		forward_UnicomService_SendCommunication_0(annotatedContext, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)
 	})
-
-	mux.Handle("GET", pattern_Unicom_GetStatus_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
+	mux.Handle(http.MethodGet, pattern_UnicomService_GetStatus_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
 		ctx, cancel := context.WithCancel(req.Context())
 		defer cancel()
 		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
-		var err error
-		ctx, err = runtime.AnnotateContext(ctx, mux, req, "/unicom.api.v1.Unicom/GetStatus", runtime.WithHTTPPathPattern("/unicom/v1/status/{id}"))
+		annotatedContext, err := runtime.AnnotateContext(ctx, mux, req, "/unicom.api.v1.UnicomService/GetStatus", runtime.WithHTTPPathPattern("/unicom/v1/status/{id}"))
 		if err != nil {
 			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
 			return
 		}
-		resp, md, err := request_Unicom_GetStatus_0(ctx, inboundMarshaler, client, req, pathParams)
-		ctx = runtime.NewServerMetadataContext(ctx, md)
+		resp, md, err := request_UnicomService_GetStatus_0(annotatedContext, inboundMarshaler, client, req, pathParams)
+		annotatedContext = runtime.NewServerMetadataContext(annotatedContext, md)
 		if err != nil {
-			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
+			runtime.HTTPError(annotatedContext, mux, outboundMarshaler, w, req, err)
 			return
 		}
-
-		forward_Unicom_GetStatus_0(ctx, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)
-
+		forward_UnicomService_GetStatus_0(annotatedContext, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)
 	})
-
 	return nil
 }
 
 var (
-	pattern_Unicom_SendCommunication_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 2, 2}, []string{"unicom", "v1", "send-communication"}, ""))
-
-	pattern_Unicom_GetStatus_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 2, 2, 1, 0, 4, 1, 5, 3}, []string{"unicom", "v1", "status", "id"}, ""))
+	pattern_UnicomService_SendCommunication_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 2, 2}, []string{"unicom", "v1", "send-communication"}, ""))
+	pattern_UnicomService_GetStatus_0         = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 2, 2, 1, 0, 4, 1, 5, 3}, []string{"unicom", "v1", "status", "id"}, ""))
 )
 
 var (
-	forward_Unicom_SendCommunication_0 = runtime.ForwardResponseMessage
-
-	forward_Unicom_GetStatus_0 = runtime.ForwardResponseMessage
+	forward_UnicomService_SendCommunication_0 = runtime.ForwardResponseMessage
+	forward_UnicomService_GetStatus_0         = runtime.ForwardResponseMessage
 )
