@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.2.0
+// - protoc-gen-go-grpc v1.5.1
 // - protoc             (unknown)
 // source: unicom/api/v1/service.proto
 
@@ -15,191 +15,173 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
-// UnicomClient is the client API for Unicom service.
+const (
+	UnicomService_SendCommunication_FullMethodName   = "/unicom.api.v1.UnicomService/SendCommunication"
+	UnicomService_StreamCommunication_FullMethodName = "/unicom.api.v1.UnicomService/StreamCommunication"
+	UnicomService_GetStatus_FullMethodName           = "/unicom.api.v1.UnicomService/GetStatus"
+)
+
+// UnicomServiceClient is the client API for UnicomService service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
-type UnicomClient interface {
+type UnicomServiceClient interface {
 	SendCommunication(ctx context.Context, in *SendCommunicationRequest, opts ...grpc.CallOption) (*SendResponse, error)
-	StreamCommunication(ctx context.Context, opts ...grpc.CallOption) (Unicom_StreamCommunicationClient, error)
+	StreamCommunication(ctx context.Context, opts ...grpc.CallOption) (grpc.BidiStreamingClient[StreamCommunicationRequest, SendResponse], error)
 	GetStatus(ctx context.Context, in *GetStatusRequest, opts ...grpc.CallOption) (*GetStatusResponse, error)
 }
 
-type unicomClient struct {
+type unicomServiceClient struct {
 	cc grpc.ClientConnInterface
 }
 
-func NewUnicomClient(cc grpc.ClientConnInterface) UnicomClient {
-	return &unicomClient{cc}
+func NewUnicomServiceClient(cc grpc.ClientConnInterface) UnicomServiceClient {
+	return &unicomServiceClient{cc}
 }
 
-func (c *unicomClient) SendCommunication(ctx context.Context, in *SendCommunicationRequest, opts ...grpc.CallOption) (*SendResponse, error) {
+func (c *unicomServiceClient) SendCommunication(ctx context.Context, in *SendCommunicationRequest, opts ...grpc.CallOption) (*SendResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(SendResponse)
-	err := c.cc.Invoke(ctx, "/unicom.api.v1.Unicom/SendCommunication", in, out, opts...)
+	err := c.cc.Invoke(ctx, UnicomService_SendCommunication_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *unicomClient) StreamCommunication(ctx context.Context, opts ...grpc.CallOption) (Unicom_StreamCommunicationClient, error) {
-	stream, err := c.cc.NewStream(ctx, &Unicom_ServiceDesc.Streams[0], "/unicom.api.v1.Unicom/StreamCommunication", opts...)
+func (c *unicomServiceClient) StreamCommunication(ctx context.Context, opts ...grpc.CallOption) (grpc.BidiStreamingClient[StreamCommunicationRequest, SendResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &UnicomService_ServiceDesc.Streams[0], UnicomService_StreamCommunication_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &unicomStreamCommunicationClient{stream}
+	x := &grpc.GenericClientStream[StreamCommunicationRequest, SendResponse]{ClientStream: stream}
 	return x, nil
 }
 
-type Unicom_StreamCommunicationClient interface {
-	Send(*StreamCommunicationRequest) error
-	Recv() (*SendResponse, error)
-	grpc.ClientStream
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type UnicomService_StreamCommunicationClient = grpc.BidiStreamingClient[StreamCommunicationRequest, SendResponse]
 
-type unicomStreamCommunicationClient struct {
-	grpc.ClientStream
-}
-
-func (x *unicomStreamCommunicationClient) Send(m *StreamCommunicationRequest) error {
-	return x.ClientStream.SendMsg(m)
-}
-
-func (x *unicomStreamCommunicationClient) Recv() (*SendResponse, error) {
-	m := new(SendResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
-
-func (c *unicomClient) GetStatus(ctx context.Context, in *GetStatusRequest, opts ...grpc.CallOption) (*GetStatusResponse, error) {
+func (c *unicomServiceClient) GetStatus(ctx context.Context, in *GetStatusRequest, opts ...grpc.CallOption) (*GetStatusResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(GetStatusResponse)
-	err := c.cc.Invoke(ctx, "/unicom.api.v1.Unicom/GetStatus", in, out, opts...)
+	err := c.cc.Invoke(ctx, UnicomService_GetStatus_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-// UnicomServer is the server API for Unicom service.
-// All implementations should embed UnimplementedUnicomServer
-// for forward compatibility
-type UnicomServer interface {
+// UnicomServiceServer is the server API for UnicomService service.
+// All implementations should embed UnimplementedUnicomServiceServer
+// for forward compatibility.
+type UnicomServiceServer interface {
 	SendCommunication(context.Context, *SendCommunicationRequest) (*SendResponse, error)
-	StreamCommunication(Unicom_StreamCommunicationServer) error
+	StreamCommunication(grpc.BidiStreamingServer[StreamCommunicationRequest, SendResponse]) error
 	GetStatus(context.Context, *GetStatusRequest) (*GetStatusResponse, error)
 }
 
-// UnimplementedUnicomServer should be embedded to have forward compatible implementations.
-type UnimplementedUnicomServer struct {
-}
+// UnimplementedUnicomServiceServer should be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedUnicomServiceServer struct{}
 
-func (UnimplementedUnicomServer) SendCommunication(context.Context, *SendCommunicationRequest) (*SendResponse, error) {
+func (UnimplementedUnicomServiceServer) SendCommunication(context.Context, *SendCommunicationRequest) (*SendResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method SendCommunication not implemented")
 }
-func (UnimplementedUnicomServer) StreamCommunication(Unicom_StreamCommunicationServer) error {
+func (UnimplementedUnicomServiceServer) StreamCommunication(grpc.BidiStreamingServer[StreamCommunicationRequest, SendResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method StreamCommunication not implemented")
 }
-func (UnimplementedUnicomServer) GetStatus(context.Context, *GetStatusRequest) (*GetStatusResponse, error) {
+func (UnimplementedUnicomServiceServer) GetStatus(context.Context, *GetStatusRequest) (*GetStatusResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetStatus not implemented")
 }
+func (UnimplementedUnicomServiceServer) testEmbeddedByValue() {}
 
-// UnsafeUnicomServer may be embedded to opt out of forward compatibility for this service.
-// Use of this interface is not recommended, as added methods to UnicomServer will
+// UnsafeUnicomServiceServer may be embedded to opt out of forward compatibility for this service.
+// Use of this interface is not recommended, as added methods to UnicomServiceServer will
 // result in compilation errors.
-type UnsafeUnicomServer interface {
-	mustEmbedUnimplementedUnicomServer()
+type UnsafeUnicomServiceServer interface {
+	mustEmbedUnimplementedUnicomServiceServer()
 }
 
-func RegisterUnicomServer(s grpc.ServiceRegistrar, srv UnicomServer) {
-	s.RegisterService(&Unicom_ServiceDesc, srv)
+func RegisterUnicomServiceServer(s grpc.ServiceRegistrar, srv UnicomServiceServer) {
+	// If the following call pancis, it indicates UnimplementedUnicomServiceServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
+	s.RegisterService(&UnicomService_ServiceDesc, srv)
 }
 
-func _Unicom_SendCommunication_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+func _UnicomService_SendCommunication_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(SendCommunicationRequest)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
 	if interceptor == nil {
-		return srv.(UnicomServer).SendCommunication(ctx, in)
+		return srv.(UnicomServiceServer).SendCommunication(ctx, in)
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/unicom.api.v1.Unicom/SendCommunication",
+		FullMethod: UnicomService_SendCommunication_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(UnicomServer).SendCommunication(ctx, req.(*SendCommunicationRequest))
+		return srv.(UnicomServiceServer).SendCommunication(ctx, req.(*SendCommunicationRequest))
 	}
 	return interceptor(ctx, in, info, handler)
 }
 
-func _Unicom_StreamCommunication_Handler(srv interface{}, stream grpc.ServerStream) error {
-	return srv.(UnicomServer).StreamCommunication(&unicomStreamCommunicationServer{stream})
+func _UnicomService_StreamCommunication_Handler(srv interface{}, stream grpc.ServerStream) error {
+	return srv.(UnicomServiceServer).StreamCommunication(&grpc.GenericServerStream[StreamCommunicationRequest, SendResponse]{ServerStream: stream})
 }
 
-type Unicom_StreamCommunicationServer interface {
-	Send(*SendResponse) error
-	Recv() (*StreamCommunicationRequest, error)
-	grpc.ServerStream
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type UnicomService_StreamCommunicationServer = grpc.BidiStreamingServer[StreamCommunicationRequest, SendResponse]
 
-type unicomStreamCommunicationServer struct {
-	grpc.ServerStream
-}
-
-func (x *unicomStreamCommunicationServer) Send(m *SendResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
-
-func (x *unicomStreamCommunicationServer) Recv() (*StreamCommunicationRequest, error) {
-	m := new(StreamCommunicationRequest)
-	if err := x.ServerStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
-
-func _Unicom_GetStatus_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+func _UnicomService_GetStatus_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(GetStatusRequest)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
 	if interceptor == nil {
-		return srv.(UnicomServer).GetStatus(ctx, in)
+		return srv.(UnicomServiceServer).GetStatus(ctx, in)
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/unicom.api.v1.Unicom/GetStatus",
+		FullMethod: UnicomService_GetStatus_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(UnicomServer).GetStatus(ctx, req.(*GetStatusRequest))
+		return srv.(UnicomServiceServer).GetStatus(ctx, req.(*GetStatusRequest))
 	}
 	return interceptor(ctx, in, info, handler)
 }
 
-// Unicom_ServiceDesc is the grpc.ServiceDesc for Unicom service.
+// UnicomService_ServiceDesc is the grpc.ServiceDesc for UnicomService service.
 // It's only intended for direct use with grpc.RegisterService,
 // and not to be introspected or modified (even as a copy)
-var Unicom_ServiceDesc = grpc.ServiceDesc{
-	ServiceName: "unicom.api.v1.Unicom",
-	HandlerType: (*UnicomServer)(nil),
+var UnicomService_ServiceDesc = grpc.ServiceDesc{
+	ServiceName: "unicom.api.v1.UnicomService",
+	HandlerType: (*UnicomServiceServer)(nil),
 	Methods: []grpc.MethodDesc{
 		{
 			MethodName: "SendCommunication",
-			Handler:    _Unicom_SendCommunication_Handler,
+			Handler:    _UnicomService_SendCommunication_Handler,
 		},
 		{
 			MethodName: "GetStatus",
-			Handler:    _Unicom_GetStatus_Handler,
+			Handler:    _UnicomService_GetStatus_Handler,
 		},
 	},
 	Streams: []grpc.StreamDesc{
 		{
 			StreamName:    "StreamCommunication",
-			Handler:       _Unicom_StreamCommunication_Handler,
+			Handler:       _UnicomService_StreamCommunication_Handler,
 			ServerStreams: true,
 			ClientStreams: true,
 		},
